requirements.txt
asyncio
aiohttp
khl.py
json